// src/utils/interfaces/Driver.ts
import { Photo } from "./Photo";
import { Vehicle } from "./Vehicle";
import { VehicleOwner } from "./VehicleOwner";
import { Wallet } from "./Wallet";
import { DriverStatus } from "@/utils/enums/driverStatusEnums";
import { StatusHistoryItem } from "./StatusHistory";
import { RequestedFreightItem } from "./RequestedFreightItem";
import { OwnerData } from "./VehicleOwnerData";
import { DriverAttachments } from "./DriverAttachments";
export interface Driver {
	_id: string;
	active: boolean;
	cnh: string;
	cnhPhoto: Photo;
	cpf: string;
	creationDate: string;
	updateDate: string;
	id: string;
	name: string;
	password: string;
	rgPhoto: Photo;
	phoneNumber: string;
	phoneNumberSecondary: string;
	proofResidencePhoto: Photo;
	requestedFreights: RequestedFreightItem[];
	status: DriverStatus;
	notificationToken: string[];
	statusHistory: StatusHistoryItem[];
	userPhoto: Photo;
	vehicle: Vehicle;
	vehicleOwner: VehicleOwner;
	wallet: Wallet;
	validated: boolean;
	acknowledge: boolean;
	updateAcknowledge: boolean;
	hashPassword: string;
	favoriteFreights: string[];
	city: string;
	state: string;
	email?: string;
	owner: OwnerData;
	attachments: DriverAttachments;
}